(core) general, shmpage: bracketed nested IFs to shut clang sanitizer up (core) audio: dropped extraneous object counter (code) video: changed tick counting to silence incorrect integer overflow warning (frameserver) net: added disconnect command, switched IDs to use a random XOR cookie (frameserver) decode: bad arg fix (frameserver) libretro: integer overflow in asamples boundary safeguard (lua) replanned net_ state transfer functions, removed old ones (lua) event_net_connected missing ID fix (lua) added argument type check for recordset definition (gridle) started reworking layering for status overlay icons to be shown properly (disconnect, network died etc.)

+++ arcan_video.h
@@ -25,7 +25,7 @@
 #ifndef RGBAPACK
 #define RGBAPACK(r, g, b, a, d)\
 {\
-	((uint32_t *)(d))[0] = (a << 24) | (b << 16) | (g << 8) | r;\
+	((uint32_t *)(d))[0] = ((uint32_t)a << 24) | ((uint32_t)b << 16) | ((uint32_t)g << 8) | (uint32_t)r;\
 }
 #endif
